requirements.txt
fastapi[standard]==0.129.0
pydantic==2.12.5
uvicorn[standard]==0.40.0
ulid-py==1.1.0
httpx==0.28.1
